Support different return types
